Implemented home screen
Also included a lot of other changes that came with the addition of a
lot of ugly code segments that will need to be scrubbed.  Generated a
static class (MeasureMe) to be initialized upon the creation of the
activity which holds if there is user data and/or what that user data
is.